fix and cleanup
