+++ src/vs/code/electron-sandbox/issue/issueReporterServiceNew.ts
@@ -1,0 +1,508 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+import { $, reset, windowOpenNoOpener } from 'vs/base/browser/dom';
+import { mainWindow } from 'vs/base/browser/window';
+import { Codicon } from 'vs/base/common/codicons';
+import { groupBy } from 'vs/base/common/collections';
+import { CancellationError } from 'vs/base/common/errors';
+import { isMacintosh } from 'vs/base/common/platform';
+import { ThemeIcon } from 'vs/base/common/themables';
+import { URI } from 'vs/base/common/uri';
+import { IssueReporterData as IssueReporterModelData } from 'vs/code/browser/issue/issueReporterModel';
+import { BaseIssueReporterService, hide, show } from 'vs/code/browser/issue/issue';
+import { localize } from 'vs/nls';
+import { isRemoteDiagnosticError } from 'vs/platform/diagnostics/common/diagnostics';
+import { IIssueMainService, IssueReporterData, IssueReporterExtensionData, IssueReporterWindowConfiguration, IssueType } from 'vs/platform/issue/common/issue';
+import { INativeHostService } from 'vs/platform/native/common/native';
+import { applyZoom, zoomIn, zoomOut } from 'vs/platform/window/electron-sandbox/window';
+
+// GitHub has let us know that we could up our limit here to 8k. We chose 7500 to play it safe.
+// ref https://github.com/microsoft/vscode/issues/159191
+const MAX_URL_LENGTH = 7500;
+
+
+export class IssueReporterNew extends BaseIssueReporterService {
+	constructor(
+		private readonly configuration: IssueReporterWindowConfiguration,
+		@INativeHostService private readonly nativeHostService: INativeHostService,
+		@IIssueMainService issueMainService: IIssueMainService
+	) {
+		super(configuration.disableExtensions, configuration.data, configuration.os, configuration.product, mainWindow, false, issueMainService);
+
+		this.issueMainService.$getSystemInfo().then(info => {
+			this.issueReporterModel.update({ systemInfo: info });
+			this.receivedSystemInfo = true;
+
+			this.updateSystemInfo(this.issueReporterModel.getData());
+			this.updatePreviewButtonState();
+		});
+		if (configuration.data.issueType === IssueType.PerformanceIssue) {
+			this.issueMainService.$getPerformanceInfo().then(info => {
+				this.updatePerformanceInfo(info as Partial<IssueReporterData>);
+			});
+		}
+
+		this.setEventHandlers();
+		applyZoom(configuration.data.zoomLevel, mainWindow);
+		this.handleExtensionData(configuration.data.enabledExtensions);
+		this.updateExperimentsInfo(configuration.data.experiments);
+		this.updateRestrictedMode(configuration.data.restrictedMode);
+		this.updateUnsupportedMode(configuration.data.isUnsupported);
+	}
+
+	private handleExtensionData(extensions: IssueReporterExtensionData[]) {
+		const installedExtensions = extensions.filter(x => !x.isBuiltin);
+		const { nonThemes, themes } = groupBy(installedExtensions, ext => {
+			return ext.isTheme ? 'themes' : 'nonThemes';
+		});
+
+		const numberOfThemeExtesions = themes && themes.length;
+		this.issueReporterModel.update({ numberOfThemeExtesions, enabledNonThemeExtesions: nonThemes, allExtensions: installedExtensions });
+		this.updateExtensionTable(nonThemes, numberOfThemeExtesions);
+		if (this.disableExtensions || installedExtensions.length === 0) {
+			(<HTMLButtonElement>this.getElementById('disableExtensions')).disabled = true;
+		}
+
+		this.updateExtensionSelector(installedExtensions);
+	}
+
+	private async sendReporterMenu(extension: IssueReporterExtensionData): Promise<IssueReporterData | undefined> {
+		try {
+			const data = await this.issueMainService.$sendReporterMenu(extension.id, extension.name);
+			return data;
+		} catch (e) {
+			console.error(e);
+			return undefined;
+		}
+	}
+
+	public override setEventHandlers(): void {
+		super.setEventHandlers();
+
+		// Keep all event listerns involving window and issue creation
+		this.previewButton.onDidClick(async () => {
+			this.delayedSubmit.trigger(async () => {
+				this.createIssue();
+			});
+		});
+
+		this.addEventListener('disableExtensions', 'click', () => {
+			this.issueMainService.$reloadWithExtensionsDisabled();
+		});
+
+		this.addEventListener('extensionBugsLink', 'click', (e: Event) => {
+			const url = (<HTMLElement>e.target).innerText;
+			windowOpenNoOpener(url);
+		});
+
+		this.addEventListener('disableExtensions', 'keydown', (e: Event) => {
+			e.stopPropagation();
+			if ((e as KeyboardEvent).keyCode === 13 || (e as KeyboardEvent).keyCode === 32) {
+				this.issueMainService.$reloadWithExtensionsDisabled();
+			}
+		});
+
+
+		// THIS IS THE MAIN IMPORTANT PART
+		mainWindow.document.onkeydown = async (e: KeyboardEvent) => {
+			const cmdOrCtrlKey = isMacintosh ? e.metaKey : e.ctrlKey;
+			// Cmd/Ctrl+Enter previews issue and closes window
+			if (cmdOrCtrlKey && e.keyCode === 13) {
+				this.delayedSubmit.trigger(async () => {
+					if (await this.createIssue()) {
+						this.close();
+					}
+				});
+			}
+
+			// Cmd/Ctrl + w closes issue window
+			if (cmdOrCtrlKey && e.keyCode === 87) {
+				e.stopPropagation();
+				e.preventDefault();
+
+				const issueTitle = (<HTMLInputElement>this.getElementById('issue-title'))!.value;
+				const { issueDescription } = this.issueReporterModel.getData();
+				if (!this.hasBeenSubmitted && (issueTitle || issueDescription)) {
+					// fire and forget
+					this.issueMainService.$showConfirmCloseDialog();
+				} else {
+					this.close();
+				}
+			}
+
+			// Cmd/Ctrl + zooms in
+			if (cmdOrCtrlKey && e.keyCode === 187) {
+				zoomIn(mainWindow);
+			}
+
+			// Cmd/Ctrl - zooms out
+			if (cmdOrCtrlKey && e.keyCode === 189) {
+				zoomOut(mainWindow);
+			}
+
+			// With latest electron upgrade, cmd+a is no longer propagating correctly for inputs in this window on mac
+			// Manually perform the selection
+			if (isMacintosh) {
+				if (cmdOrCtrlKey && e.keyCode === 65 && e.target) {
+					if (e.target instanceof HTMLInputElement || e.target instanceof HTMLTextAreaElement) {
+						(<HTMLInputElement>e.target).select();
+					}
+				}
+			}
+		};
+	}
+
+	public override async submitToGitHub(issueTitle: string, issueBody: string, gitHubDetails: { owner: string; repositoryName: string }): Promise<boolean> {
+		const url = `https://api.github.com/repos/${gitHubDetails.owner}/${gitHubDetails.repositoryName}/issues`;
+		const init = {
+			method: 'POST',
+			body: JSON.stringify({
+				title: issueTitle,
+				body: issueBody
+			}),
+			headers: new Headers({
+				'Content-Type': 'application/json',
+				'Authorization': `Bearer ${this.data.githubAccessToken}`
+			})
+		};
+
+		const response = await fetch(url, init);
+		if (!response.ok) {
+			console.error('Invalid GitHub URL provided.');
+			return false;
+		}
+		const result = await response.json();
+		await this.nativeHostService.openExternal(result.html_url);
+		this.close();
+		return true;
+	}
+
+	public override async createIssue(): Promise<boolean> {
+		const selectedExtension = this.issueReporterModel.getData().selectedExtension;
+		const hasUri = this.nonGitHubIssueUrl;
+		// Short circuit if the extension provides a custom issue handler
+		if (hasUri) {
+			const url = this.getExtensionBugsUrl();
+			if (url) {
+				this.hasBeenSubmitted = true;
+				await this.nativeHostService.openExternal(url);
+				return true;
+			}
+		}
+
+		if (!this.validateInputs()) {
+			// If inputs are invalid, set focus to the first one and add listeners on them
+			// to detect further changes
+			const invalidInput = mainWindow.document.getElementsByClassName('invalid-input');
+			if (invalidInput.length) {
+				(<HTMLInputElement>invalidInput[0]).focus();
+			}
+
+			this.addEventListener('issue-title', 'input', _ => {
+				this.validateInput('issue-title');
+			});
+
+			this.addEventListener('description', 'input', _ => {
+				this.validateInput('description');
+			});
+
+			this.addEventListener('issue-source', 'change', _ => {
+				this.validateInput('issue-source');
+			});
+
+			if (this.issueReporterModel.fileOnExtension()) {
+				this.addEventListener('extension-selector', 'change', _ => {
+					this.validateInput('extension-selector');
+				});
+			}
+
+			return false;
+		}
+
+		this.hasBeenSubmitted = true;
+
+		const issueTitle = (<HTMLInputElement>this.getElementById('issue-title')).value;
+		const issueBody = this.issueReporterModel.serialize();
+
+		let issueUrl = this.getIssueUrl();
+		if (!issueUrl) {
+			console.error('No issue url found');
+			return false;
+		}
+
+		if (selectedExtension?.uri) {
+			const uri = URI.revive(selectedExtension.uri);
+			issueUrl = uri.toString();
+		}
+
+		const gitHubDetails = this.parseGitHubUrl(issueUrl);
+		if (this.data.githubAccessToken && gitHubDetails) {
+			return this.submitToGitHub(issueTitle, issueBody, gitHubDetails);
+		}
+
+		const baseUrl = this.getIssueUrlWithTitle((<HTMLInputElement>this.getElementById('issue-title')).value, issueUrl);
+		let url = baseUrl + `&body=${encodeURIComponent(issueBody)}`;
+
+		if (url.length > MAX_URL_LENGTH) {
+			try {
+				url = await this.writeToClipboard(baseUrl, issueBody);
+			} catch (_) {
+				console.error('Writing to clipboard failed');
+				return false;
+			}
+		}
+
+		await this.nativeHostService.openExternal(url);
+		return true;
+	}
+
+	public override async writeToClipboard(baseUrl: string, issueBody: string): Promise<string> {
+		const shouldWrite = await this.issueMainService.$showClipboardDialog();
+		if (!shouldWrite) {
+			throw new CancellationError();
+		}
+
+		await this.nativeHostService.writeClipboardText(issueBody);
+
+		return baseUrl + `&body=${encodeURIComponent(localize('pasteData', "We have written the needed data into your clipboard because it was too large to send. Please paste."))}`;
+	}
+
+	private updateSystemInfo(state: IssueReporterModelData) {
+		const target = mainWindow.document.querySelector<HTMLElement>('.block-system .block-info');
+
+		if (target) {
+			const systemInfo = state.systemInfo!;
+			const renderedDataTable = $('table', undefined,
+				$('tr', undefined,
+					$('td', undefined, 'CPUs'),
+					$('td', undefined, systemInfo.cpus || '')
+				),
+				$('tr', undefined,
+					$('td', undefined, 'GPU Status' as string),
+					$('td', undefined, Object.keys(systemInfo.gpuStatus).map(key => `${key}: ${systemInfo.gpuStatus[key]}`).join('\n'))
+				),
+				$('tr', undefined,
+					$('td', undefined, 'Load (avg)' as string),
+					$('td', undefined, systemInfo.load || '')
+				),
+				$('tr', undefined,
+					$('td', undefined, 'Memory (System)' as string),
+					$('td', undefined, systemInfo.memory)
+				),
+				$('tr', undefined,
+					$('td', undefined, 'Process Argv' as string),
+					$('td', undefined, systemInfo.processArgs)
+				),
+				$('tr', undefined,
+					$('td', undefined, 'Screen Reader' as string),
+					$('td', undefined, systemInfo.screenReader)
+				),
+				$('tr', undefined,
+					$('td', undefined, 'VM'),
+					$('td', undefined, systemInfo.vmHint)
+				)
+			);
+			reset(target, renderedDataTable);
+
+			systemInfo.remoteData.forEach(remote => {
+				target.appendChild($<HTMLHRElement>('hr'));
+				if (isRemoteDiagnosticError(remote)) {
+					const remoteDataTable = $('table', undefined,
+						$('tr', undefined,
+							$('td', undefined, 'Remote'),
+							$('td', undefined, remote.hostName)
+						),
+						$('tr', undefined,
+							$('td', undefined, ''),
+							$('td', undefined, remote.errorMessage)
+						)
+					);
+					target.appendChild(remoteDataTable);
+				} else {
+					const remoteDataTable = $('table', undefined,
+						$('tr', undefined,
+							$('td', undefined, 'Remote'),
+							$('td', undefined, remote.latency ? `${remote.hostName} (latency: ${remote.latency.current.toFixed(2)}ms last, ${remote.latency.average.toFixed(2)}ms average)` : remote.hostName)
+						),
+						$('tr', undefined,
+							$('td', undefined, 'OS'),
+							$('td', undefined, remote.machineInfo.os)
+						),
+						$('tr', undefined,
+							$('td', undefined, 'CPUs'),
+							$('td', undefined, remote.machineInfo.cpus || '')
+						),
+						$('tr', undefined,
+							$('td', undefined, 'Memory (System)' as string),
+							$('td', undefined, remote.machineInfo.memory)
+						),
+						$('tr', undefined,
+							$('td', undefined, 'VM'),
+							$('td', undefined, remote.machineInfo.vmHint)
+						)
+					);
+					target.appendChild(remoteDataTable);
+				}
+			});
+		}
+	}
+
+	public updateExtensionSelector(extensions: IssueReporterExtensionData[]): void {
+		interface IOption {
+			name: string;
+			id: string;
+		}
+
+		const extensionOptions: IOption[] = extensions.map(extension => {
+			return {
+				name: extension.displayName || extension.name || '',
+				id: extension.id
+			};
+		});
+
+		// Sort extensions by name
+		extensionOptions.sort((a, b) => {
+			const aName = a.name.toLowerCase();
+			const bName = b.name.toLowerCase();
+			if (aName > bName) {
+				return 1;
+			}
+
+			if (aName < bName) {
+				return -1;
+			}
+
+			return 0;
+		});
+
+		const makeOption = (extension: IOption, selectedExtension?: IssueReporterExtensionData): HTMLOptionElement => {
+			const selected = selectedExtension && extension.id === selectedExtension.id;
+			return $<HTMLOptionElement>('option', {
+				'value': extension.id,
+				'selected': selected || ''
+			}, extension.name);
+		};
+
+		const extensionsSelector = this.getElementById<HTMLSelectElement>('extension-selector');
+		if (extensionsSelector) {
+			const { selectedExtension } = this.issueReporterModel.getData();
+			reset(extensionsSelector, this.makeOption('', localize('selectExtension', "Select extension"), true), ...extensionOptions.map(extension => makeOption(extension, selectedExtension)));
+
+			if (!selectedExtension) {
+				extensionsSelector.selectedIndex = 0;
+			}
+
+			this.addEventListener('extension-selector', 'change', async (e: Event) => {
+				this.clearExtensionData();
+				const selectedExtensionId = (<HTMLInputElement>e.target).value;
+				this.selectedExtension = selectedExtensionId;
+				const extensions = this.issueReporterModel.getData().allExtensions;
+				const matches = extensions.filter(extension => extension.id === selectedExtensionId);
+				if (matches.length) {
+					this.issueReporterModel.update({ selectedExtension: matches[0] });
+					const selectedExtension = this.issueReporterModel.getData().selectedExtension;
+					if (selectedExtension) {
+						const iconElement = document.createElement('span');
+						iconElement.classList.add(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
+						this.setLoading(iconElement);
+						const openReporterData = await this.sendReporterMenu(selectedExtension);
+						if (openReporterData) {
+							if (this.selectedExtension === selectedExtensionId) {
+								this.removeLoading(iconElement, true);
+								this.configuration.data = openReporterData;
+								this.data = openReporterData;
+							} else if (this.selectedExtension !== selectedExtensionId) {
+							}
+						}
+						else {
+							if (!this.loadingExtensionData) {
+								iconElement.classList.remove(...ThemeIcon.asClassNameArray(Codicon.loading), 'codicon-modifier-spin');
+							}
+							this.removeLoading(iconElement);
+							// if not using command, should have no configuration data in fields we care about and check later.
+							this.clearExtensionData();
+
+							// case when previous extension was opened from normal openIssueReporter command
+							selectedExtension.data = undefined;
+							selectedExtension.uri = undefined;
+						}
+						if (this.selectedExtension === selectedExtensionId) {
+							// repopulates the fields with the new data given the selected extension.
+							this.updateExtensionStatus(matches[0]);
+							this.openReporter = false;
+						}
+					} else {
+						this.issueReporterModel.update({ selectedExtension: undefined });
+						this.clearSearchResults();
+						this.clearExtensionData();
+						this.validateSelectedExtension();
+						this.updateExtensionStatus(matches[0]);
+					}
+				}
+			});
+		}
+
+		this.addEventListener('problem-source', 'change', (_) => {
+			this.validateSelectedExtension();
+		});
+	}
+
+	public override setLoading(element: HTMLElement) {
+		// Show loading
+		this.openReporter = true;
+		this.loadingExtensionData = true;
+		this.updatePreviewButtonState();
+
+		const extensionDataCaption = this.getElementById('extension-id')!;
+		hide(extensionDataCaption);
+
+		const extensionDataCaption2 = Array.from(mainWindow.document.querySelectorAll('.ext-parens'));
+		extensionDataCaption2.forEach(extensionDataCaption2 => hide(extensionDataCaption2));
+
+		const showLoading = this.getElementById('ext-loading')!;
+		show(showLoading);
+		while (showLoading.firstChild) {
+			showLoading.removeChild(showLoading.firstChild);
+		}
+		showLoading.append(element);
+
+		this.renderBlocks();
+	}
+
+	public override removeLoading(element: HTMLElement, fromReporter: boolean = false) {
+		this.openReporter = fromReporter;
+		this.loadingExtensionData = false;
+		this.updatePreviewButtonState();
+
+		const extensionDataCaption = this.getElementById('extension-id')!;
+		show(extensionDataCaption);
+
+		const extensionDataCaption2 = Array.from(mainWindow.document.querySelectorAll('.ext-parens'));
+		extensionDataCaption2.forEach(extensionDataCaption2 => show(extensionDataCaption2));
+
+		const hideLoading = this.getElementById('ext-loading')!;
+		hide(hideLoading);
+		if (hideLoading.firstChild) {
+			hideLoading.removeChild(element);
+		}
+		this.renderBlocks();
+	}
+
+	private updateRestrictedMode(restrictedMode: boolean) {
+		this.issueReporterModel.update({ restrictedMode });
+	}
+
+	private updateUnsupportedMode(isUnsupported: boolean) {
+		this.issueReporterModel.update({ isUnsupported });
+	}
+
+	private updateExperimentsInfo(experimentInfo: string | undefined) {
+		this.issueReporterModel.update({ experimentInfo });
+		const target = mainWindow.document.querySelector<HTMLElement>('.block-experiments .block-info');
+		if (target) {
+			target.textContent = experimentInfo ? experimentInfo : localize('noCurrentExperiments', "No current experiments.");
+		}
+	}
+}
